// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/profile-overview-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DietologistRelationshipHttpResponse } from './dietologist-relationship-http-response';
import { NotificationPreferencesHttpResponse } from './notification-preferences-http-response';
import { UserHttpResponse } from './user-http-response';
import { WebPushSubscriptionHttpResponse } from './web-push-subscription-http-response';

export interface ProfileOverviewHttpResponse {
    dietologistRelationship?: DietologistRelationshipHttpResponse | null;
    notificationPreferences?: NotificationPreferencesHttpResponse;
    user?: UserHttpResponse;
    webPushSubscriptions?: Array<WebPushSubscriptionHttpResponse>;
}