// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/update-user-http-request.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DashboardLayoutHttpModel } from './dashboard-layout-http-model';

export interface UpdateUserHttpRequest {
    activityLevel?: string | null;
    birthDate?: string | null;
    dashboardLayout?: DashboardLayoutHttpModel | null;
    fastingPushNotificationsEnabled?: boolean | null;
    firstName?: string | null;
    gender?: string | null;
    heightCm?: number | null;
    hydrationGoal?: number | null;
    isActive?: boolean | null;
    language?: string | null;
    lastName?: string | null;
    profileImage?: string | null;
    profileImageAssetId?: string | null;
    pushNotificationsEnabled?: boolean | null;
    socialPushNotificationsEnabled?: boolean | null;
    stepGoal?: number | null;
    theme?: string | null;
    timeZoneId?: string | null;
    uiStyle?: string | null;
    username?: string | null;
    weightKg?: number | null;
}