// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/user-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DashboardLayoutHttpModel } from './dashboard-layout-http-model';

export interface UserHttpResponse {
    activityLevel?: string;
    aiConsentAcceptedAt?: string | null;
    birthDate?: string | null;
    carbTarget?: number | null;
    dailyCalorieTarget?: number | null;
    dashboardLayout?: DashboardLayoutHttpModel | null;
    desiredWaistCm?: number | null;
    desiredWeightKg?: number | null;
    email?: string | null;
    fastingCheckInFollowUpReminderHours?: number;
    fastingCheckInReminderHours?: number;
    fastingPushNotificationsEnabled?: boolean;
    fatTarget?: number | null;
    fiberTarget?: number | null;
    firstName?: string | null;
    gender?: string | null;
    hasGoogleIdentity?: boolean;
    hasPassword?: boolean;
    hasTelegramIdentity?: boolean;
    heightCm?: number | null;
    hydrationGoal?: number | null;
    id?: string;
    isActive?: boolean;
    isEmailConfirmed?: boolean;
    language?: string | null;
    lastLoginAtUtc?: string | null;
    lastName?: string | null;
    mustChangePassword?: boolean;
    profileImage?: string | null;
    profileImageAssetId?: string | null;
    proteinTarget?: number | null;
    pushNotificationsEnabled?: boolean;
    socialPushNotificationsEnabled?: boolean;
    stepGoal?: number | null;
    surfaceStyle?: string;
    theme?: string | null;
    timeZoneId?: string | null;
    uiStyle?: string | null;
    username?: string | null;
    waterGoal?: number | null;
    weightKg?: number | null;
}