// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/waist-goal-history-page-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WaistGoalHistoryHttpResponse } from './waist-goal-history-http-response';

export interface WaistGoalHistoryPageHttpResponse {
    items?: Array<WaistGoalHistoryHttpResponse>;
    nextCursor?: string | null;
}