// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/weight-goal-history-page-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WeightGoalHistoryHttpResponse } from './weight-goal-history-http-response';

export interface WeightGoalHistoryPageHttpResponse {
    items?: Array<WeightGoalHistoryHttpResponse>;
    nextCursor?: string | null;
}